NEW - task 40: Really large attr-values result in ArrayOutOfBounce in File-backend https://projects.uni-konstanz.de/issues/show/40
adapted ByteBufferSourceAndSink and JavaCrytpoImpl to use a variable size of bytes. Test cases work

git-svn-id: https://svn.uni-konstanz.de/disy/Treetank/trunk/treetank@5149 e3ddb328-5bfe-0310-b762-aafcbcbd2528

+++ src/main/java/com/treetank/utils/CryptoJavaImpl.java
@@ -80,7 +80,7 @@ public class CryptoJavaImpl implements ICrypto {
         for (final byte byteVal : result) {
             buffer.writeByte(byteVal);
         }
-        return (short) (buffer.position());
+        return buffer.position();
     }
 
     /**
@@ -118,7 +118,7 @@ public class CryptoJavaImpl implements ICrypto {
         for (final byte byteVal : result) {
             buffer.writeByte(byteVal);
         }
-        return (short) (result.length + 24);
+        return result.length + 24;
     }
 
 }
